chore(ui): fix circular dep
