+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/form-manipulation.ts
@@ -247,75 +247,6 @@ export const getAllowedDropRegions = (form: BuilderForm, element: FormElement): 
 };
 
 /**
- * Validates the structure of a form.
- *
- * The form structure is valid if:
- * - The root element is a container
- * - Starting from the root element, all children referenced are reachable
- * - There are no extra elements in the form that are not reachable from the root element
- * - The root element has no parentId and is a container
- * - Non-root elements have a parentId
- * - All parent elements are containers
- * - All elements with a parentId are children of their parent
- *
- * @param form The form to validate
- *
- * @returns True if the form structure is valid, false otherwise
- */
-export const validateFormStructure = (form: BuilderForm): boolean => {
-  const { elements, rootElementId } = form;
-
-  const rootElement = elements[rootElementId];
-  const isRootElementAContainer = rootElement !== undefined && isContainerElement(rootElement);
-
-  const childrenFoundInTree = new Set<string>();
-
-  const findChildren = (elementId: string): boolean => {
-    const element = elements[elementId];
-    if (!element) {
-      // Element not found
-      return false;
-    }
-    childrenFoundInTree.add(elementId);
-    if (element.id === rootElementId) {
-      // Special handling for root
-      if (element.parentId !== undefined) {
-        // Root element must not have a parent
-        return false;
-      }
-    } else {
-      // Handling for all other elements
-      if (element.parentId === undefined) {
-        // Element must have a parent
-        return false;
-      }
-      const parent = elements[element.parentId];
-      if (!parent) {
-        // Parent must exist
-        return false;
-      }
-      if (!isContainerElement(parent)) {
-        // Parent must be a container
-        return false;
-      }
-      if (!parent.data.children.includes(elementId)) {
-        // Element must be a child of its parent
-        return false;
-      }
-    }
-    if (isContainerElement(element) && element.data.children.length > 0) {
-      return element.data.children.every(findChildren);
-    }
-    return true;
-  };
-
-  const noMissingChildren = findChildren(rootElementId);
-  const noExtraElements = Object.keys(elements).length === childrenFoundInTree.size;
-
-  return isRootElementAContainer && noMissingChildren && noExtraElements;
-};
-
-/**
  * Checks if a form is empty.
  * A form is empty if it only contains the root element and the root element has no children.
  * @param form The form to check
